Fix ULTRA weird bug w/ pickling and singledispatch
